// src/app/models/commentaire.ts
export interface Commentaire {
    com_cde_id: number;
    com_id: number;
    com_message: string;
    com_image: string;
    com_livraison_avec_soin: boolean;
    com_au_dela_attentes: boolean;
    com_articles_conformes: boolean;
    com_respect_instructions: boolean;
    com_dans_delais: boolean;
    com_articles_qualite: boolean;
}